pyproject.toml
[build-system]
requires = ["hatchling>=1.20"]
build-backend = "hatchling.build"

[project]
name = "hefesto-dualsense4unix"
version = "3.0.0"
description = "Hefesto - Dualsense4Unix: Linux adaptive trigger daemon for DualSense controllers"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ["dualsense", "ps5", "adaptive-triggers", "linux", "gamepad", "dualsense4unix"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pydualsense>=0.7.5",
    "textual>=0.47",
    "typer>=0.9",
    "pydantic>=2.0",
    "python-xlib>=0.33",
    "evdev>=1.6",
    "structlog>=23.0",
    "rich>=13.0",
    "platformdirs>=4.0",
    "filelock>=3.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.7",
    "types-python-xlib",
]
emulation = [
    "python-uinput>=0.11",
]
tray = [
    "PyGObject>=3.44",
]
# FEAT-METRICS-01 — opcional; text exposition manual não precisa desta dep.
# Instalar se quiser usar prometheus_client diretamente (ex: dashboards avançados).
metrics = [
    "prometheus-client>=0.19",
]

[project.scripts]
hefesto-dualsense4unix = "hefesto_dualsense4unix.cli.app:main"
hefesto-dualsense4unix-gui = "hefesto_dualsense4unix.app.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/hefesto_dualsense4unix"]
include = ["src/hefesto_dualsense4unix/gui/*.glade"]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = ["tests/fixtures/**"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "pydualsense",
    "pydualsense.*",
    "textual",
    "textual.*",
    "evdev",
    "Xlib",
    "Xlib.*",
    "uinput",
    "gi",
    "gi.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
